Re-organized package using simpler structure
