+++ vincentvanbot/frompicture/predict.py
@@ -1,33 +1,10 @@
 import os
-from sklearn.neighbors import NearestNeighbors
 import joblib
-import pandas as pd
 from google.cloud import storage
+import pandas as pd
 from vincentvanbot.params import BUCKET_NAME, BUCKET_INITIAL_DATASET_FOLDER
-from vincentvanbot.preprocessing.utils import get_jpg_link
+from vincentvanbot.utils import get_jpg_link
 
-
-def train_model(df_transformed):
-    """Takes preprocessed train data as df. Returns fitted KNN model 
-    and train data image indexes (used then to refer back to initial database).
-    Saves locally model and indexes."""
-    knn_model = NearestNeighbors().fit(df_transformed)
-    
-    joblib.dump(knn_model,'model.joblib')
-    joblib.dump(df_transformed.index,'train_indexes.joblib')
-
-def save_model_to_cloud(rm=False):
-    """Uploads fitted model and related indexes to GCloud."""
-    client = storage.Client().bucket(BUCKET_NAME)
-    
-    for filename in ['model.joblib','train_indexes.joblib']:
-        storage_location = f"predict/{filename}"
-        blob = client.blob(storage_location)
-        blob.upload_from_filename(filename)
-        print(f"=> {filename} uploaded to bucket {BUCKET_NAME} inside {storage_location}")
-    if rm:
-        os.remove('model.joblib')
-        os.remove('train_indexes.joblib')
 
 def get_closest_images_indexes(user_input_transformed, nsimilar=3, rm=True):
     """Takes user_input_transformed as np.array. Downloads fitted knn model and related indexes.
@@ -82,15 +59,10 @@ def get_info_from_index(indexes, all_info=False):
 
 
 if __name__=='__main__':
-    from vincentvanbot.preprocessing.utils import preprocess_image
-    # from vincentvanbot.data import get_joblib_data
-    path = os.path.join(os.path.dirname(__file__),'..','notebooks','example-input.jpg')
+    from vincentvanbot.preprocessing import preprocess_image
+    path = os.path.join(os.path.dirname(__file__),'..','..','notebooks','example-input.jpg')
     # print(path)
     user_img = preprocess_image(path,dim=(100,100))
-    # train_df = get_joblib_data()
-
-    # train_model(train_df)
-    # save_model_to_cloud(rm=True)
     indexes = get_closest_images_indexes(user_img)
     urls = get_info_from_index(indexes)
     print(urls)
